refactor: convert all pages to client components with direct Convex hooks
- Remove preloadQuery pattern that was causing authentication issues
- Convert upcoming, event, explore pages to client components using useQuery
- Convert UserInfo to client component with conditional queries
- Delete all intermediate client wrapper components
- All pages now have real-time updates via Convex hooks

This simplifies the codebase significantly and avoids the complexity of
server-side authentication with preloadQuery.
